requirements.txt
langchain==0.0.332
flask
flask[async]
pypdf2
python-dotenv
faiss-cpu
openai
tiktoken
-e .